Source changes for 2014.0.5

+++ Source/Vision/Runtime/EnginePlugins/VisionEnginePlugin/GUI/Controls/VControlHelperObjects.cpp
@@ -2,7 +2,7 @@
  *
  * Confidential Information of Telekinesys Research Limited (t/a Havok). Not for disclosure or distribution without Havok's
  * prior written consent. This software contains code, techniques and know-how which is confidential and proprietary to Havok.
- * Product and Trade Secret source code contains trade secrets of Havok. Havok Software (C) Copyright 1999-2013 Telekinesys Research Limited t/a Havok. All Rights Reserved. Use of this software is subject to the terms of an end user license agreement.
+ * Product and Trade Secret source code contains trade secrets of Havok. Havok Software (C) Copyright 1999-2014 Telekinesys Research Limited t/a Havok. All Rights Reserved. Use of this software is subject to the terms of an end user license agreement.
  *
  */
 
@@ -258,7 +258,7 @@ void VTextState::Paint(VGraphicsInfo *pGraphics, VWindowBase *pParentWnd, VColor
         if (*szPtr == '\n')
         {
           VStaticString<512> line;
-          line.Set(szCurrentLine, szPtr - szCurrentLine);
+          line.Set(szCurrentLine, static_cast<int>(szPtr - szCurrentLine));
 
           m_lines.Add(line.AsChar());
           szCurrentLine = szPtr + 1;
@@ -300,7 +300,7 @@ void VTextState::Paint(VGraphicsInfo *pGraphics, VWindowBase *pParentWnd, VColor
           else if(strlen(szCurrentLine) <= iWrapOffset)
           {
             // End of text occurs before automatic text wrap
-            iByteOffsetAtWrapPosition = strlen(szCurrentLine);
+            iByteOffsetAtWrapPosition = static_cast<int>(strlen(szCurrentLine));
             iByteOffsetAfterWrapPosition = iByteOffsetAtWrapPosition;
           }
           else
@@ -675,9 +675,9 @@ void VTextStates::SerializeX( VArchive &ar )
 }
 
 /*
- * Havok SDK - Base file, BUILD(#20131218)
+ * Havok SDK - Base file, BUILD(#20140327)
  * 
- * Confidential Information of Havok.  (C) Copyright 1999-2013
+ * Confidential Information of Havok.  (C) Copyright 1999-2014
  * Telekinesys Research Limited t/a Havok. All Rights Reserved. The Havok
  * Logo, and the Havok buzzsaw logo are trademarks of Havok.  Title, ownership
  * rights, and intellectual property rights in the Havok software remain in
